+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Versions
 
+- [1.1.0](#110)
 - [1.0.0](#100)
 
 <br>
@@ -9,6 +10,14 @@
 ---
 
 ## Changes
+
+<br>
+
+### 1.1.0
+
+- `readonly` support for `Rater` component
+- `Review` components
+- Clean-up
 
 <br>